requirements-win.txt
pywebview==5.2
